refact: isolating separate component
