+++ src/types/index.ts
@@ -1,2 +1,3 @@
 export * from "./aid";
 export * from "./acdc";
+export * from "./report";
